Multiple changes and bug fixes on contacts and tasks

+++ app/src/main/java/fry/oldschool/adapter/TaskAdapter.java
@@ -5,6 +5,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.graphics.Color;
 import android.view.LayoutInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
@@ -12,12 +13,14 @@ import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.ImageButton;
 import android.widget.LinearLayout;
+import android.widget.PopupMenu;
 import android.widget.TextView;
 
 import java.util.ArrayList;
 
 import fry.oldschool.R;
 import fry.oldschool.activity.TaskCreateActivity;
+import fry.oldschool.utils.App;
 import fry.oldschool.utils.TaskList;
 
 public class TaskAdapter extends ArrayAdapter<TaskList>{
@@ -57,19 +60,13 @@ public class TaskAdapter extends ArrayAdapter<TaskList>{
             res = convertView;
         }
 
-        final TaskList item = getItem(position);
-
         TextView header = (TextView) res.findViewById(R.id.textview_listtemplate_header);
+        final TaskList item = getItem(position);
         final String headerText = item.name;
         header.setText(headerText);
-        /*
-        if (position == 0)
-            header.setBackgroundColor(Color.rgb(20, 150, 20));
-        else
-            header.setBackgroundColor(Color.rgb(60, 60, 180));
-        */
         LinearLayout entries = (LinearLayout) res.findViewById(R.id.linearlayout_listtemplate_id);
         entries.removeAllViews(); //Change later, because this isn't the best solution
+
         for (int i=0; i<item.length(); i++){
             CheckBox cb = new CheckBox(ctx);
             String cbText = item.getTaskName(i);
@@ -78,23 +75,33 @@ public class TaskAdapter extends ArrayAdapter<TaskList>{
             entries.addView(cb);
         }
 
-        Button delete = (Button) res.findViewById(R.id.imagebutton_listtemplate_delete);
-        delete.setOnClickListener(new View.OnClickListener() {
+        final Button more = (Button) res.findViewById(R.id.button_listtemplate_more);
+        more.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                item.delete();
+                PopupMenu popupMenu = new PopupMenu(App.mContext, more);
+                popupMenu.getMenuInflater().inflate(R.menu.task_options_popup, popupMenu.getMenu());
+
+                popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
+                    @Override
+                    public boolean onMenuItemClick(MenuItem menuItem) {
+                        if (menuItem.getTitle().equals(App.mContext.getResources().getString(R.string.edit))){
+                            Intent intent = new Intent(ctx, TaskCreateActivity.class);
+                            intent.putExtra("index", position);
+                            ctx.startActivity(intent);
+                        }
+                        else if (menuItem.getTitle().equals(App.mContext.getResources().getString(R.string.delete))){
+                            item.delete();
+                            notifyDataSetChanged();
+                        }
+
+                        return false;
+                    }
+                });
+                popupMenu.show();
             }
         });
 
-        Button edit = (Button) res.findViewById(R.id.imagebutton_listtemplate_edit);
-        edit.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(ctx, TaskCreateActivity.class);
-                intent.putExtra("index", position);
-                ctx.startActivity(intent);
-            }
-        });
 
         //Set optical design elements for list item (view)
         res.setBackgroundResource(R.drawable.listview_style);
